requirements.txt
pillow
numpy
flask